requirements.txt
awsebcli==3.10.3
boto==2.48.0
botocore==1.5.82
braintree==3.25.0
cement==2.8.2
colorama==0.3.7
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.0
Django==2.0.3
django-bootstrap-datepicker-plus==1.0.0
django-bootstrap-form==3.3
django-colorfield==0.1.14
django-crispy-forms==1.7.0
django-extensions==1.9.8
django-import-export==1.0.1
django-storages==1.6.3
django-tables2==1.21.2
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
et-xmlfile==1.0.1
gunicorn==19.4.5
jdcal==1.4
jmespath==0.9.3
oauthlib==1.0.3
odfpy==1.3.6
openpyxl==2.5.4
pathspec==0.5.0
psycopg2==2.7.4
PyJWT==1.4.0
python-dateutil==2.6.1
python-social-auth==0.2.14
python3-openid==3.0.10
pytz==2018.3
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.6.1
semantic-version==2.5.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tablib==0.12.1
tabulate==0.7.5
termcolor==1.1.0
typing==3.6.4
unicodecsv==0.14.1
websocket-client==0.44.0
whitenoise==3.3.0
xlrd==1.1.0
xlwt==1.3.0
